Rename package stringprocessing to inputprocessing
