+++ src/main/java/MainWindow.java
@@ -1,6 +1,6 @@
 import ducky.Ducky;
 
-import ducky.stringprocessing.Parser;
+import ducky.inputprocessing.Parser;
 import javafx.animation.PauseTransition;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
