+++ src/main/java/ducky/task/Deadline.java
@@ -1,6 +1,6 @@
 package ducky.task;
 
-import ducky.stringprocessing.StringifyDate;
+import ducky.inputprocessing.StringifyDate;
 import java.time.LocalDateTime;
 
 /**
